refactor(service)Refactor Stripe subscription service
Updated the Stripe subscription service to increase maintainability and readability. Instead of retrieving the user and object data each time a subscription is created, these data are now retrieved when the service class is instantiated. This reduces the number of queries and keeps the code DRY. The subscription creation and update process are also now separated into distinct methods, improving the code's organization and clarity.