Refactor(subscription): Refactor code to use stripe_id instead of subscription_id
The attribute named subscription_id was incorrectly being used to find or update subscriptions. The correct attribute to use is stripe_id, the one which is recognized by the Stripe API. All instances of subscription_id have been replaced with stripe_id to prevent any future confusion or bugs related to this issue.